% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{algoEM}
\alias{algoEM}
\title{EM algorithm for probability of budburst}
\usage{
algoEM(
  data,
  temp.params = list(temp.min.cu = -10, temp.max.cu = 15, temp.min.fu = 5, temp.max.fu
    = 35),
  init.params = list(a.cu = -2, b.cu = 5, a.fu = 15, b.fu = 4, mu = 1500, s = 750),
  origin.date = "09-01",
  control = list(proposal = "AdGl", mcsize = 5, saem.size1 = 10, saem.size2 = 0)
)
}
\arguments{
\item{data}{a list containing the plant and temperatures data}

\item{temp.params}{a list with the fixed parameters giving the minimum and maximum temperatures for computing chilling and forcing units}

\item{init.params}{a list with the initial values for the parameters to be estimated}

\item{origin.date}{the date to be used as the origin when computing cumulative sum of temperatures}

\item{control}{a list of options for the algorithm}
}
\description{
HELP PAGE TO WRITE -- EM algorithme for estimation of the parameters when they vary accross years
}
