% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chillingUnits.R
\name{chillingUnits}
\alias{chillingUnits}
\alias{chilling}
\alias{CU}
\title{Chilling units}
\usage{
chillingUnits(temp.data, var.names = list(temp = "temp", date = "date"),
  temp.min = -5, temp.max = 15, mu, s)
}
\arguments{
\item{temp.data}{a data frame with the dates and temperatures to accumulate}

\item{var.names}{the name of the temperature and date variables, in the format \code{list(temp="temp.name",date="date.name")}}

\item{temp.min}{the threshold above which the plant is accumulating chilling units}

\item{temp.max}{the threshold below which the plant is accumulating chilling units}

\item{mu}{the temperature at which the plant accumulates the highest number of chilling units}

\item{s}{the sample size of the underlying Beta distribution}
}
\value{
a vector with the chilling units corresponding to each input temperature
}
\description{
Compute chilling units accumulated by a plant, according to a beta-shaped function.
}
\details{
The accumulation of chilling units is expressed as the Beta law probability distribution function (pdf). The underlying Beta law is re-parametrized
using the mode and the so-called sample size. More precisely, using the following definition for the pdf of the Beta law with parameters \code{a} and \code{b}:
\deqn{\frac{\Gamma(a+b)}{\Gamma (a)+\Gamma (b)} \ x^{a-1}(1-x)^{1-b}}
for \eqn{0 \leq x \leq 1}, \eqn{a>0} and \eqn{b >0}. Now, we further assume here that \eqn{a > 1} and \eqn{b > 1}, so that the mode exists, and we parametrized
the Beta distribution using \eqn{\mu} and \eqn{s} defined as:
\deqn{\mu = \frac{a-1}{a+b-2}, \quad s = a+b}

Since the Beta law has its support on \eqn{[0,1]}, it is evaluated at \code{(x-temp.min)/(temp.max-temp.min)}, where \code{x} is the recorded temperature, and
\code{mu} is transformed into \code{(mu-temp.min)/(temp.max-temp.min)} to lie between 0 and 1
}
