% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dormancy_scout.R
\name{compute_dormancy_potential}
\alias{compute_dormancy_potential}
\title{Compute dormancy potential at a grid point}
\usage{
compute_dormancy_potential(x, y, x_center, y_center, method)
}
\description{
Compute dormancy potential at a grid point
}
\keyword{internal}
