% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dormancy_depth.R
\name{compute_energy_depth}
\alias{compute_energy_depth}
\title{Compute energy-based depth}
\usage{
compute_energy_depth(x, y, pattern, data)
}
\description{
Compute energy-based depth
}
\keyword{internal}
