% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dormancy_depth.R
\name{compute_entropy_depth}
\alias{compute_entropy_depth}
\title{Compute entropy-based depth}
\usage{
compute_entropy_depth(x, y, pattern, data)
}
\description{
Compute entropy-based depth
}
\keyword{internal}
