% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dormancy_risk.R
\name{compute_impact_magnitude}
\alias{compute_impact_magnitude}
\title{Compute impact magnitude}
\usage{
compute_impact_magnitude(x, y, pattern, data)
}
\description{
Compute impact magnitude
}
\keyword{internal}
