% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dormancy_depth.R
\name{compute_stability_depth}
\alias{compute_stability_depth}
\title{Compute stability-based depth}
\usage{
compute_stability_depth(x, y, pattern, data)
}
\description{
Compute stability-based depth
}
\keyword{internal}
