% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dormancy_detect.R
\name{detect_cascade_dormancy}
\alias{detect_cascade_dormancy}
\title{Detect cascade-ready dormancy}
\usage{
detect_cascade_dormancy(x, y, v1, v2, num_data, min_cluster, overall_cor)
}
\description{
Detect cascade-ready dormancy
}
\keyword{internal}
