% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dormancy_detect.R
\name{detect_conditional_dormancy}
\alias{detect_conditional_dormancy}
\title{Detect conditional dormancy}
\usage{
detect_conditional_dormancy(
  x,
  y,
  v1,
  v2,
  data,
  numeric_cols,
  n_bins,
  min_cluster,
  overall_cor
)
}
\description{
Detect conditional dormancy
}
\keyword{internal}
