% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dormancy_detect.R
\name{detect_phase_dormancy}
\alias{detect_phase_dormancy}
\title{Detect phase-based dormancy}
\usage{
detect_phase_dormancy(x, y, v1, v2, n_bins, min_cluster, overall_cor)
}
\description{
Detect phase-based dormancy
}
\keyword{internal}
