% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dormancy_detect.R
\name{detect_self_conditional}
\alias{detect_self_conditional}
\title{Detect self-conditional dormancy (when no other variables available)}
\usage{
detect_self_conditional(x, y, v1, v2, n_bins, min_cluster, overall_cor)
}
\description{
Detect self-conditional dormancy (when no other variables available)
}
\keyword{internal}
