% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dormancy_detect.R
\name{detect_threshold_dormancy}
\alias{detect_threshold_dormancy}
\title{Detect threshold-based dormancy}
\usage{
detect_threshold_dormancy(x, y, v1, v2, n_bins, min_cluster, overall_cor)
}
\description{
Detect threshold-based dormancy
}
\keyword{internal}
