% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dormancy_risk.R
\name{determine_impact_direction}
\alias{determine_impact_direction}
\title{Determine impact direction}
\usage{
determine_impact_direction(x, y, pattern, data)
}
\description{
Determine impact direction
}
\keyword{internal}
