% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dormancy_trigger.R
\name{determine_trigger_type}
\alias{determine_trigger_type}
\title{Determine the type of trigger}
\usage{
determine_trigger_type(pattern, method)
}
\description{
Determine the type of trigger
}
\keyword{internal}
