% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTemp.R
\name{extractTemp}
\alias{extractTemp}
\title{Temperatures per plant}
\usage{
extractTemp(temp.outside, temp.inside, var.names.out, var.names.in,
  data.deb, var.names.deb)
}
\arguments{
\item{temp.outside}{a data frame with the dates and temperatures outside the greenhouse}

\item{temp.inside}{a data frame with the dates and temperatures inside the greenhouse}

\item{var.names.out}{the names of the temperature and date variables, in the format \code{list(temp="temp.name",date="date.name")} for \code{temp.outside}}

\item{var.names.in}{the names of the temperature and date variables, in the format \code{list(temp="temp.name",date="date.name")} for \code{temp.inside}}

\item{data.deb}{a data frame with the information about budburst}

\item{var.names.deb}{the names of the variables representing the session, the plant ID, the harvesting date and the budburst status in \code{data.deb},
in the format \code{list(session="sess.name",plant="plant.name",harv="harv.name",budburst="bud.name")}}
}
\value{
a data frame with the accumulated temperatures outside and inside the greenhouse by the plant from 1st of january to 31st of december
}
\description{
Extract the temperatures experienced by each plant, taking into account both periods before and after harvest
}
