% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_entropy_2d}
\alias{fast_entropy_2d}
\title{Fast 2D Entropy}
\usage{
fast_entropy_2d(x, y, n_bins)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}

\item{n_bins}{Number of bins per dimension}
}
\value{
Normalized entropy value
}
\description{
Compute entropy of a 2D binned distribution.
}
\keyword{internal}
