% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_mutual_info}
\alias{fast_mutual_info}
\title{Fast Mutual Information}
\usage{
fast_mutual_info(x, y, n_bins)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}

\item{n_bins}{Number of bins for discretization}
}
\value{
Mutual information estimate
}
\description{
Estimate mutual information between two variables.
}
\keyword{internal}
