% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcingUnits.R
\name{forcingUnits}
\alias{forcingUnits}
\alias{forcing}
\alias{CU}
\title{Forcing units}
\usage{
forcingUnits(
  temp.data,
  var.names = list(temp = "temp", date = "date", duration = "duration"),
  temp.min = 5,
  temp.max = 30,
  a,
  b
)
}
\arguments{
\item{temp.data}{a data frame with the dates and temperatures to accumulate}

\item{var.names}{the name of the temperature and date variables, in the format \code{list(temp="temp.name",date="date.name",duration="duration.name")}}

\item{temp.min}{the threshold above which the plant is accumulating chilling units}

\item{temp.max}{the threshold below which the plant is accumulating chilling units}

\item{a}{the temperature at which the plant accumulates half the maximum of forcing units}

\item{b}{the rate of forcing units accumulation}
}
\value{
a vector with the forcing units corresponding to each input temperature
}
\description{
Compute forcing units accumulated by a plant, according to a beta-shaped function.
}
\details{
The accumulation of forcing units is expressed through a logistic curve with parameters \code{a} and \code{b}:
\deqn{fu(t) = \frac{1}{1 + \exp(-(t-a)/b)} }
if \eqn{temp.min \leq t \leq temp.max}, and 0 otherwise, and where \eqn{a} is the temperature at which the plant accumulates half the maximum of forcing units and \eqn{b} is the rate of accumulation
}
