% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dormancy_risk.R
\name{generate_risk_recommendations}
\alias{generate_risk_recommendations}
\title{Generate risk recommendations}
\usage{
generate_risk_recommendations(risk_scores, tolerance)
}
\description{
Generate risk recommendations
}
\keyword{internal}
