% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dormancy_trigger.R
\name{generate_trigger_recommendations}
\alias{generate_trigger_recommendations}
\title{Generate actionable recommendations from trigger analysis}
\usage{
generate_trigger_recommendations(triggers_df, sensitivity)
}
\description{
Generate actionable recommendations from trigger analysis
}
\keyword{internal}
