% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\docType{class}
\name{prior-class}
\alias{prior-class}
\alias{prior}
\title{Class "prior"}
\description{
An object of the \code{prior} class
}
\section{Slots}{

\describe{
\item{\code{distName}}{the name of the prior distribution}

\item{\code{hyperParams}}{a list of the distribution parameters (hyperparameters)}
}}

